refactor(actions): server actions을 도메인별 폴더로 재구조화
