+++ app/actions/competition/manage.ts
