+++ components/races/competition-register-dialog.tsx
@@ -4,7 +4,7 @@ import { useEffect, useMemo } from "react";
 import Link from "next/link";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { createCompetition } from "@/app/actions/create-competition";
+import { createCompetition } from "@/app/actions/competition/create";
 import {
   competitionRegisterSchema,
   type CompetitionRegisterValues,
