+++ components/races/race-list-view.tsx
@@ -7,8 +7,8 @@ import { ChevronDown, ChevronRight, Plus } from "lucide-react";
 import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
 import { CardItem } from "@/components/ui/card";
-import { getPastGigangCompetitions } from "@/app/actions/get-past-gigang-competitions";
-import { revalidateCompetitions } from "@/app/actions/revalidate-competitions";
+import { getPastGigangCompetitions } from "@/app/actions/competition/get-past-gigang";
+import { revalidateCompetitions } from "@/app/actions/competition/revalidate";
 import { CompetitionDetailDialog } from "./competition-detail-dialog";
 import { CompetitionRegisterDialog } from "./competition-register-dialog";
 import type { Competition, CompetitionRegistration, MemberStatus } from "./types";
